merging dev into new_flow_context-frame_copy to show conflicts

+++ disa_app/lib/view_data_records_manager.py
@@ -149,61 +149,6 @@ def manage_reference_put( rec_id: str, payload: bytes, request_user_id: int ) ->
     ## end def manage_reference_put()
 
 
-# def manage_reference_put( rec_id: str, payload: bytes, request_user_id: int ) -> dict:
-#     """ Handles api call when 'Create' button is hit in `/editor/records/?doc_id=(123)`.
-#         Called by views.data_records() """
-#     log.debug( 'starting manage_reference_put()' )
-#     session = make_session()
-#     data: dict = json.loads( payload )
-#     log.debug( f'data, ```{pprint.pformat(data)}```' )
-#     try:  # for debugging; remove afterwards
-
-#         reference_type: models_sqlalchemy.ReferenceType = get_or_create_type( data['record_type'], models_alch.ReferenceType, session )
-
-#         # ref = models.Reference.query.get(refId)
-#         rfrnc = session.query( models_alch.Reference ).get( rec_id )
-
-#         rfrnc.locations = []
-#         rfrnc = process_record_locations( data['locations'], rfrnc, session )
-
-#         try:
-#             rfrnc.date = datetime.datetime.strptime(data['date'], '%m/%d/%Y')
-#         except:
-#             rfrnc.date = None
-#         rfrnc.reference_type_id = reference_type.id
-#         rfrnc.national_context_id = data['national_context']
-#         rfrnc.transcription = data['transcription']
-#         session.add( rfrnc )
-#         session.commit()
-
-#         stamp_edit( request_user_id, rfrnc, session )
-
-#         data = { 'rec': {} }
-#         data['rec']['id'] = rfrnc.id
-#         data['rec']['date'] = ''
-#         if rfrnc.date:
-#             data['rec']['date'] = '{}/{}/{}'.format(rfrnc.date.month,
-#                 rfrnc.date.day, rfrnc.date.year)
-#         data['rec']['citation'] = rfrnc.citation.id
-#         data['rec']['transcription'] = rfrnc.transcription
-#         data['rec']['national_context'] = rfrnc.national_context_id
-#         data['rec']['locations'] = [
-#             { 'label':l.location.name, 'value':l.location.name,
-#                 'id': l.location.id } for l in rfrnc.locations ]
-#         data['rec']['record_type'] = {'label': rfrnc.reference_type.name,
-#             'value': rfrnc.reference_type.name, 'id':rfrnc.reference_type.id }
-
-#         # context =  { 'redirect': reverse( 'edit_record_url', kwargs={'rec_id': rfrnc.id} ) }
-#         log.debug( f'data, ```{data}```' )
-#     except:
-#         log.exception( '\n\nexception...' )
-#         raise Exception( 'problem; see logs' )
-
-#     return data
-
-#     ## end def manage_reference_put()
-
-
 def manage_post( payload: bytes, request_user_id: int ) -> dict:
     """ Handles api call when 'Create' button is hit in `/editor/records/?doc_id=(123)`.
         Called by views.data_records() """
@@ -252,7 +197,7 @@ def manage_post( payload: bytes, request_user_id: int ) -> dict:
     return context
 
 
-def manage_reference_delete( rfrnc_id: str ) -> HttpResponseRedirect:  # or, much less likely, HttpResponseNotFound
+def manage_reference_delete( rfrnc_id: str ) -> dict:  # or, much less likely, HttpResponseNotFound
     """ Handles api call when red `x` button is clicked...
         ...and then the 'Confirm delete' button is clicked in, eg, <http://127.0.0.1:8000/editor/documents/(123)/>.
         Called by views.data_reference()
@@ -260,23 +205,37 @@ def manage_reference_delete( rfrnc_id: str ) -> HttpResponseRedirect:  # or, muc
               ...but a good TODO would be to refactor that view and have a general views.data_record() url...
               ...handle the full CRUD set of methods -- which would all call this data_records_manager.py file """
     log.debug( 'starting manage_delete()' )
-    assert type(rfrnc_id) == str
-    session = make_session()
-    existing = session.query( models_alch.Reference ).get( rfrnc_id )
-    if existing:
-        log.debug( 'found reference to delete' )
-        cite = existing.citation  # why did I get this?
-        session.delete( existing )
-        session.commit()
-    else:
-        log.debug( 'TODO: return a 404 or bad-request' )
-        pass
-    redirect_url = reverse( 'edit_citation_url', kwargs={'cite_id': cite.id} )
-    context =  { 'redirect': redirect_url }
+    context = {}
+    try:
+        assert type(rfrnc_id) == str
+        if rfrnc_id == 'undefined':
+            context = { 'err': '400 / Bad Request' }
+            log.warning( f'should not receive `undefined` rfrnc_id' )
+    except Exception as e:
+        log.exception( 'Bad reference-id.' )
+        context = { 'err': '400 / Bad Request' }
+        log.warning( f'should not receive a rfrnc_id of, ``{rfrnc_id}``' )
+    if context == {}:
+        try:
+            session = make_session()
+            existing = session.query( models_alch.Reference ).get( rfrnc_id )
+            if existing:
+                log.debug( 'found reference to delete' )
+                cite = existing.citation  # why did I get this?
+                session.delete( existing )
+                session.commit()
+                redirect_url = reverse( 'edit_citation_url', kwargs={'cite_id': cite.id} )
+                context =  { 'redirect': redirect_url }
+            else:
+                context = { 'err': '404 / Not Found' }
+                log.warning( f'should not receive a not-found rfrnc_id of, ``{rfrnc_id}``' )
+        except Exception as e:
+            log.exception( 'Problem deleting reference.' )
     log.debug( f'context, ``{context}``' )
     return context
 
-# def manage_reference_delete( rfrnc_id: str ) -> HttpResponseRedirect:  # or, much less likely, HttpResponseNotFound
+
+# def manage_reference_delete( rfrnc_id: str ) -> dict:  # or, much less likely, HttpResponseNotFound
 #     """ Handles api call when red `x` button is clicked...
 #         ...and then the 'Confirm delete' button is clicked in, eg, <http://127.0.0.1:8000/editor/documents/(123)/>.
 #         Called by views.data_reference()
@@ -284,19 +243,21 @@ def manage_reference_delete( rfrnc_id: str ) -> HttpResponseRedirect:  # or, muc
 #               ...but a good TODO would be to refactor that view and have a general views.data_record() url...
 #               ...handle the full CRUD set of methods -- which would all call this data_records_manager.py file """
 #     log.debug( 'starting manage_delete()' )
+#     assert type(rfrnc_id) == str
 #     session = make_session()
 #     existing = session.query( models_alch.Reference ).get( rfrnc_id )
 #     if existing:
-#         cite = existing.citation
+#         log.debug( 'found reference to delete' )
+#         cite = existing.citation  # why did I get this?
 #         session.delete( existing )
 #         session.commit()
-#         redirect_url = reverse( 'edit_citation_url', kwargs={'cite_id': cite.id} )
-#         log.debug( f'redirect_url, ```{redirect_url}```' )
-#         rsp = HttpResponseRedirect( redirect_url )  # if this doesn't work, i can just hit this url with requests and return the output.
 #     else:
-#         rsp = HttpResponseNotFound
-#     log.debug( f'returning rsp, ```{rsp}```' )
-#     return rsp
+#         log.debug( 'TODO: return a 404 or bad-request' )
+#         pass
+#     redirect_url = reverse( 'edit_citation_url', kwargs={'cite_id': cite.id} )
+#     context =  { 'redirect': redirect_url }
+#     log.debug( f'context, ``{context}``' )
+#     return context
 
 
 # -------------
